refactor(web): isolate v2 compat projection
