+++ docs/ops/audit/web_contract_v2_frontend_architecture_guard_report.md
@@ -2,12 +2,8 @@
 
 - mode: debt_lock
 - status: PASS
-- findings: 5
+- findings: 0
 
 ## Findings
 
-- [P1] v2_to_legacy_projection_entry `frontend/apps/web/src/api/contract.ts` count=2/2
-- [P1] legacy_field_descriptor_synthesis `frontend/apps/web/src/api/contract.ts` count=2/2
-- [P1] legacy_form_layout_synthesis `frontend/apps/web/src/api/contract.ts` count=2/2
-- [P1] legacy_subview_policy_synthesis `frontend/apps/web/src/api/contract.ts` count=2/2
-- [P1] legacy_layout_button_synthesis `frontend/apps/web/src/api/contract.ts` count=2/2
+- none
