+++ docs/ops/iterations/delivery_context_switch_log_v1.md
@@ -32829,3 +32829,4 @@ Legacy compliance note: `/api/scenes/my` is deprecated; successor endpoint is `/
 - web_contract_v2_action_view_data_boundary_batch_c21: `timestamp=2026-05-13T18:42:36+08:00。Layer Target=Frontend ActionView Data Boundary；Module=frontend/apps/web/src/views/ActionView.vue + frontend/apps/web/src/app/runtime/actionViewDataRuntime.ts + scripts/verify/web_contract_v2_frontend_architecture_guard.py；Reason=ActionView 不应直接 import api/data，先把 list/write/delete/batch/favorite 数据操作收束到 runtime 边界，同时保持既有 list runtime hook 输入契约不变。完成：新增 actionViewDataRuntime 导出 ActionView 专用 list/write/delete/batch/favorite operations；ActionView 改为消费 runtime 并通过显式参数名映射传入既有 hooks；architecture guard 将 action_view_direct_data_imports 锁定为 0，债务 finding 从 9 降至 8。验证：pnpm -C frontend/apps/web typecheck PASS；python3 -m py_compile scripts/verify/web_contract_v2_frontend_architecture_guard.py PASS；python3 scripts/verify/web_contract_v2_frontend_architecture_guard.py PASS，findings=8；make verify.unified_page_contract.v2 PASS；git diff --check PASS；pnpm -C frontend/apps/web lint PASS；make verify.docs.all PASS；ENV=dev ENV_FILE=.env.dev make verify.frontend.build PASS；FRONTEND_PROFILE=daily make frontend.restart PASS；FRONTEND_URL=http://127.0.0.1:5174 DB_NAME=sc_demo E2E_LOGIN=wutao E2E_PASSWORD=123456 RECORD_ID=908 make verify.unified_page_contract.v2.web_form_shadow_browser.host PASS，artifact=artifacts/playwright/web-contract-v2-form-shadow/20260513T104546。branch=codex/web-contract-v2-frontend-upgrade short_sha=45dc72a9。`
 - web_contract_v2_contract_form_data_boundary_batch_c22: `timestamp=2026-05-13T18:50:24+08:00。Layer Target=Frontend ContractForm Data Boundary；Module=frontend/apps/web/src/pages/ContractFormPage.vue + frontend/apps/web/src/app/runtime/contractFormDataRuntime.ts + scripts/verify/web_contract_v2_frontend_architecture_guard.py；Reason=ContractFormPage 不应直接 import api/data，先把 form read/write/create 与 relation option list/create 下沉到 runtime 边界，后续再接 ContractV2Runtime declared dataSource/action operations。完成：新增 contractFormDataRuntime 导出 ContractForm 专用 read/list/create/write operations；ContractFormPage 改为消费 runtime；architecture guard 将 product_page_direct_data_imports 锁定为 0，债务 finding 从 8 降至 7。验证：pnpm -C frontend/apps/web typecheck PASS；python3 -m py_compile scripts/verify/web_contract_v2_frontend_architecture_guard.py PASS；python3 scripts/verify/web_contract_v2_frontend_architecture_guard.py PASS，findings=7；make verify.unified_page_contract.v2 PASS；git diff --check PASS；pnpm -C frontend/apps/web lint PASS；make verify.docs.all PASS；ENV=dev ENV_FILE=.env.dev make verify.frontend.build PASS；FRONTEND_PROFILE=daily make frontend.restart PASS；FRONTEND_URL=http://127.0.0.1:5174 DB_NAME=sc_demo E2E_LOGIN=wutao E2E_PASSWORD=123456 RECORD_ID=908 make verify.unified_page_contract.v2.web_form_shadow_browser.host PASS，artifact=artifacts/playwright/web-contract-v2-form-shadow/20260513T105214。branch=codex/web-contract-v2-frontend-upgrade short_sha=b5b0dc95。`
 - web_contract_v2_lite_fallback_compat_boundary_batch_c23: `timestamp=2026-05-13T18:55:34+08:00。Layer Target=Frontend Lite Preview Compat Boundary；Module=frontend/apps/web/src/app/contracts/unifiedPageContractLite.ts + frontend/apps/web/src/app/contracts/unifiedPageContractLiteCompat.ts + frontend/apps/web/src/api/contract.ts + scripts/verify/web_contract_v2_frontend_architecture_guard.py；Reason=Lite preview 的 legacy_default fallback 是兼容协议，不应散落在 API adapter 与 Lite schema 默认文件中。完成：新增 unifiedPageContractLiteCompat 导出 LITE_PREVIEW_LEGACY_FALLBACK_MODE、LitePreviewFallbackMode 与判定函数；LitePreviewEnvelope 与 extractLitePreviewEnvelope 改为消费 compat boundary；loadModelLitePreviewContract 改为使用 compat 常量；architecture guard 将 lite_legacy_default_fallback 与 api_contract_lite_legacy_default_fallback 锁定为 0，债务 finding 从 7 降至 5。验证：pnpm -C frontend/apps/web typecheck PASS；python3 -m py_compile scripts/verify/web_contract_v2_frontend_architecture_guard.py PASS；python3 scripts/verify/web_contract_v2_frontend_architecture_guard.py PASS，findings=5；make verify.unified_page_contract.v2 PASS；git diff --check PASS；pnpm -C frontend/apps/web lint PASS；make verify.docs.all PASS；ENV=dev ENV_FILE=.env.dev make verify.frontend.build PASS；FRONTEND_PROFILE=daily make frontend.restart PASS；FRONTEND_URL=http://127.0.0.1:5174 DB_NAME=sc_demo E2E_LOGIN=wutao E2E_PASSWORD=123456 RECORD_ID=908 make verify.unified_page_contract.v2.web_form_shadow_browser.host PASS，artifact=artifacts/playwright/web-contract-v2-form-shadow/20260513T110302。branch=codex/web-contract-v2-frontend-upgrade short_sha=40753898。`
+- web_contract_v2_compat_projection_boundary_batch_c24: `timestamp=2026-05-13T19:07:52+08:00。Layer Target=Frontend V2 Compat Projection Boundary；Module=frontend/apps/web/src/api/contract.ts + frontend/apps/web/src/app/runtime/unifiedPageContractV2CompatProjection.ts + scripts/verify/web_contract_v2_frontend_architecture_guard.py；Reason=ui.contract.v2 到旧 ActionContract runtime shape 的投影属于兼容层，不应继续驻留在 API adapter 默认文件。完成：新增 unifiedPageContractV2CompatProjection 承载 buildRuntimeProjectionFromV2、buildLegacyFieldDescriptor、buildLegacyFormLayout、buildLegacySubViews、collectV2LayoutButtons 与 adaptUnifiedPageContractV2Raw；api/contract.ts 改为只负责 intent 请求、错误重写与调用 compat projection boundary；architecture guard 将剩余 5 个 api/contract.ts legacy projection rules 锁定为 0，债务 finding 从 5 降至 0。验证：pnpm -C frontend/apps/web typecheck PASS；python3 -m py_compile scripts/verify/web_contract_v2_frontend_architecture_guard.py PASS；python3 scripts/verify/web_contract_v2_frontend_architecture_guard.py PASS，findings=0；make verify.unified_page_contract.v2 PASS；git diff --check PASS；pnpm -C frontend/apps/web lint PASS；make verify.docs.all PASS；ENV=dev ENV_FILE=.env.dev make verify.frontend.build PASS；FRONTEND_PROFILE=daily make frontend.restart PASS；FRONTEND_URL=http://127.0.0.1:5174 DB_NAME=sc_demo E2E_LOGIN=wutao E2E_PASSWORD=123456 RECORD_ID=908 make verify.unified_page_contract.v2.web_form_shadow_browser.host PASS，artifact=artifacts/playwright/web-contract-v2-form-shadow/20260513T111022。branch=codex/web-contract-v2-frontend-upgrade short_sha=be988b64。`
