+++ scripts/verify/web_contract_v2_frontend_architecture_guard.py
@@ -43,7 +43,7 @@ DEBT_RULES: tuple[DebtRule, ...] = (
         severity="P1",
         path="api/contract.ts",
         pattern="buildRuntimeProjectionFromV2",
-        max_count=2,
+        max_count=0,
         rationale="Contract v2 is still projected back into a legacy runtime shape.",
         next_action="Move v2-to-legacy projection behind an explicit compat boundary and remove default imports.",
     ),
@@ -52,7 +52,7 @@ DEBT_RULES: tuple[DebtRule, ...] = (
         severity="P1",
         path="api/contract.ts",
         pattern="buildLegacyFieldDescriptor",
-        max_count=2,
+        max_count=0,
         rationale="Frontend synthesizes legacy field descriptors from v2 widgets.",
         next_action="Consume typed v2 widget descriptors directly from ContractV2Store.",
     ),
@@ -61,7 +61,7 @@ DEBT_RULES: tuple[DebtRule, ...] = (
         severity="P1",
         path="api/contract.ts",
         pattern="buildLegacyFormLayout",
-        max_count=2,
+        max_count=0,
         rationale="Frontend synthesizes legacy form layout from v2 widgets.",
         next_action="Render containerTree/layoutContract directly through the v2 renderer registry.",
     ),
@@ -70,7 +70,7 @@ DEBT_RULES: tuple[DebtRule, ...] = (
         severity="P1",
         path="api/contract.ts",
         pattern="buildLegacySubViews",
-        max_count=2,
+        max_count=0,
         rationale="Frontend synthesizes relation subviews and edit policies.",
         next_action="Use relation_entry/dataContract/actionRuleList declarations without policy defaults.",
     ),
@@ -79,7 +79,7 @@ DEBT_RULES: tuple[DebtRule, ...] = (
         severity="P1",
         path="api/contract.ts",
         pattern="collectV2LayoutButtons",
-        max_count=2,
+        max_count=0,
         rationale="Frontend walks v2 layout to synthesize legacy buttons.",
         next_action="Render actionContract.actionRuleList and widget status directly.",
     ),
